Security audit round 2: fix 20 vulnerabilities (2 Critical, 7 High, 6 Medium, 5 Low)
Critical fixes:
- C-5: GuardVpnService restart runnable cancellation - cancel on stopVpn()
- C-6: OnnxMlEngine session leak - close old session on model reload

High fixes:
- H-7: GuardVpnService file stream leak - close in finally block
- H-8: LocalDnsEngine @Volatile - add volatile to shared mutable fields
- H-9: LocalDnsEngine BK-tree race - double-checked locking with cacheLock
- H-10: TieredAlertEngine rate limiter race - synchronized canLaunchAlert()
- H-11: AlertDataHolder pendingRunnables leak - cleanup on consume()
- H-12: MlEngineWithFallback state race - synchronized getState()
- H-13: DohClient certificate pinning - Cloudflare pins added

Medium fixes:
- M-9: AlertActivity TTL expiry - finish() instead of showing wrong alert
- M-11: BootReceiver DI bypass - use SharedPreferences directly
- M-13: DohClient certificate pinning - Cloudflare pins added
- M-14: Bm25Engine clearFeedbackIndex - synchronized with feedbackLock

Low fixes:
- L-1: CooldownManager integer overflow - toLong() conversion
- L-3: FeatureExtractor regex reuse - use class-level urlPattern
- L-4: AlertDataHolder clear() - remove only own callbacks

Build: SUCCESS
ADB: PASS (no crashes)

+++ app/src/main/java/com/tianshang/guard/core/alert/CooldownManager.kt
@@ -10,7 +10,8 @@ class CooldownManager(private val prefs: GuardPreferences) {
     fun isInCooldown(key: String, cooldownSeconds: Int): Boolean {
         if (cooldownSeconds <= 0) return false
         val lastTrigger = triggerTimes[key] ?: return false
-        return (System.currentTimeMillis() - lastTrigger) < cooldownSeconds * 1000
+        // L-1: Use toLong() to prevent integer overflow
+        return (System.currentTimeMillis() - lastTrigger) < cooldownSeconds.toLong() * 1000
     }
 
     fun recordTrigger(key: String) {
